% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getF}
\alias{getF}
\title{Calculates inbreeding coefficients of individuals in the pedigree.}
\usage{
getF(pedigree)
}
\arguments{
\item{pedigree}{data.frame with three columns: id, sire and dam}
}
\value{
Numeric vector
}
\description{
Calculates inbreeding coefficients of individuals in the pedigree.
}
\details{
Uses Rcpp package for calculations
}
\examples{
id = c(1,2,3,4,5,6,7,8,9,10)
sire = c(0,0,1,1,3,5,5,5,7,7)
dam = c(0,0,2,2,2,4,4,6,6,8)
ped = as.data.frame(cbind(id,sire,dam))
ped$F = getF(ped)

}

