% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getOrdPed}
\alias{getOrdPed}
\title{Orders a pedigree (parents before offspring) using Kahn's Algorithm.}
\usage{
getOrdPed(pedigree, verbose = FALSE)
}
\arguments{
\item{pedigree}{data.frame with three columns: id, sire and dam}
}
\value{
Integer vector
}
\description{
Orders a pedigree (parents before offspring) using Kahn's Algorithm.
}
\details{
Kahn's algorithm is a topological sorting method commonly applied to networks. It's only feasible for acyclic directed graphs, so any cycles (ex. an individual that has itself as a parent) will return an error and must be corrected.
}
\examples{
id = c("A","B","C","D","E","F","G","H","I","J","K")
sire = c("0","0","0","0","A","A","C","A","G","G","A")
dam = c("0","0","0","0","B","H","D","D","B","E","H")
ped = as.data.frame(cbind(id,sire,dam))
newOrd = getOrdPed(ped)
pedReord = ped[ped$id[c(newOrd)],]

}
\references{
Kahn, Arthur B. (1962), "Topological sorting of large networks", Communications of the ACM
}

