% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getOrdPed}
\alias{getOrdPed}
\title{Orders a pedigree (parents before offspring) using Kahn's Algorithm.}
\usage{
getOrdPed(pedigree)
}
\arguments{
\item{pedigree}{data.frame with three columns: id, sire and dam}
}
\value{
Character vector
}
\description{
Orders a pedigree (parents before offspring) using Kahn's Algorithm.
}
\details{
Kahn's algorithm is a topological sorting method commonly applied to networks. It's only feasible for acyclic directed graphs, so any cycles (ex. an individual that has itself as a parent) will return an error and must be corrected.
}
\examples{
id = c(1,2,3,4,5,6,7,8,9,10)
sire = c(0,0,1,1,3,5,5,5,7,7)
dam = c(0,0,2,2,2,4,4,6,6,8)
ped = as.data.frame(cbind(id,sire,dam))
newOrd = getOrdPed(ped)

}
\references{
Kahn, Arthur B. (1962), "Topological sorting of large networks", Communications of the ACM
}

